public survey route
